requirements.txt
flask==2.3.3
requests==2.32.3
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
langchain-core==0.3.45
langgraph==0.2.70
typing-extensions==4.12.2
groq==0.18.0
langchain-groq==0.2.4
httpx==0.25.2
-e